refactor(training): reduce file complexity in ActiveWorkout
Extracted state management, API calls, and business logic from ActiveWorkout.tsx into a dedicated hook (useActiveWorkout.ts).
Also moved the timer logic into its own hook (useElapsed.ts).
This change leaves the component to focus on rendering the UI, significantly improving readability and maintainability.
